Separating Groovy Template engine
Separating the groovy template engine so that a Play 1 and Play 2 specific implementation can be done.
Most of the work is done via abstractions and there is now a TemplateEngine that needs to be initialized.

Utility methods (for interoperability between Play 1 and Play2) are all abstracted in the TemplateUtils class.
Concerns such as route resolution and more are handled in the TemplateEngine implementation.
There are also other version-specific implementations for e.g. FastTags and GroovyTemplate.